+++ cring.h
@@ -112,8 +112,8 @@ CRING_NAME(pop) (CRING_T() *q, CRING_TYPE *data, size_t count);
 
 
 enum cring_filestatus
-CRING_NAME(readput) (CRING_T() *q, int fd, int *count);
+CRING_NAME(readput) (CRING_T() *q, int fd, size_t *count);
 
 
 enum cring_filestatus
-CRING_NAME(popwrite) (CRING_T() *q, int fd, int *count);
+CRING_NAME(popwrite) (CRING_T() *q, int fd, size_t *count);
